// frontend/src/api/generated/model/sessionPhase.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenCode Studio API
 * API for OpenCode Studio - AI-powered development platform
 * OpenAPI spec version: 0.1.0
 */

export type SessionPhase = typeof SessionPhase[keyof typeof SessionPhase];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SessionPhase = {
  planning: 'planning',
  implementation: 'implementation',
  review: 'review',
  fix: 'fix',
} as const;