// frontend/src/api/generated/model/sessionStatus.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenCode Studio API
 * API for OpenCode Studio - AI-powered development platform
 * OpenAPI spec version: 0.1.0
 */

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SessionStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  aborted: 'aborted',
} as const;